requirements.txt
beautifulsoup4
bs4
mysql-connector-python
python-dateutil
requests
selenium
urllib3
